rename molfile to Molfile
